+++ lib/src/interpreter.dart
@@ -30,3 +30,21 @@ abstract class MeasurementInterpreter<T> {
   /// The multiplier for the configured unit.
   final double _unitMultiplier;
 }
+
+/// A [MeasurementInterpreter] that truncates its results to `int`s.
+///
+/// This is useful for measurements that should not be represented fractionally,
+/// e.g. the number of items in a collection.
+abstract class TruncatingMeasurementInterpreter<T>
+    extends MeasurementInterpreter<T> {
+  @override
+  double _of(final num value) =>
+      (value.toDouble() * _unitMultiplier).truncateToDouble();
+
+  @override
+  double _from(final num value) => value.toDouble() / _unitMultiplier;
+
+  /// Constructs a [TruncatingMeasurementInterpreter].
+  const TruncatingMeasurementInterpreter._(final double unitMultiplier)
+      : super._(unitMultiplier);
+}