+++ lib/src/prefix.dart
@@ -5,6 +5,7 @@ class MeasurementPrefix
     with
         DistancePrefix,
         MassPrefix,
+        QuantityPrefix,
         TemperaturePrefix,
         TimePrefix,
         VolumePrefix {